huggingface-quantization-pipeline/requirements.txt
torch
transformers
datasets
numpy
pandas
scipy
matplotlib
pyyaml
torchvision
onnx
onnxruntime